Separate helper for creating the default redux store for www pages
